Remove storage files and move code to mps and util
